backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
xgboost==2.0.2
cors==1.0.1
